+++ app/src/main/java/com/example/healthx/db/UserDao.kt
@@ -17,6 +17,6 @@ interface UserDao {
     fun getUserData(): LiveData<List<UserData>>
 
     @Query("UPDATE userData SET steps = :newSteps, calories = :newCalories, distance = :newDistance WHERE id = :userId")
-    suspend fun updatePedometerDetails(userId: String, newSteps: Int, newCalories: Int, newDistance: Int)
+    suspend fun updatePedometerDetails(userId: String, newSteps: Int, newCalories: Float, newDistance: Float)
 
 }
